eerly_studio/pyproject.toml
[project]
name = "eerly-studio"
version = "0.1.0"
description = "Eerly Studio Authentication Service for Aegra"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "structlog>=24.1.0",
    "python-multipart>=0.0.9",
    "alembic>=1.13.0",
    "asgi-correlation-id>=4.3.0",
    "langgraph-sdk>=0.3.4",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eerly_studio"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
